Review and re-arrangement of the WPS processes and their utilities.
