+++ vires/vires/processes/util/filters/__init__.py
@@ -1,10 +1,10 @@
 #-------------------------------------------------------------------------------
 #
-#  Process Utilities - cache backend wrapper
+#  Data filters
 #
 # Authors: Martin Paces <martin.paces@eox.at>
 #-------------------------------------------------------------------------------
-# Copyright (C) 2017 EOX IT Services GmbH
+# Copyright (C) 2016 EOX IT Services GmbH
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -24,18 +24,32 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 #-------------------------------------------------------------------------------
+# pylint: disable=missing-docstring
 
-from functools import wraps
-from eoxserver.backends.cache import setup_cache_session, shutdown_cache_session
+from numpy import empty
+from .base import Filter
+from .range import ScalarRangeFilter, VectorComponentRangeFilter
+from .bounding_box import BoundingBoxFilter
+from .temporal_sampling import MinStepSampler, GroupingSampler, ExtraSampler
 
-def with_cache_session(func):
-    """ Decorator setting up the EOxServer cache session. """
-    @wraps(func)
-    def __wrapper__(*args, **kwargs):
-        setup_cache_session()
-        try:
-            response = func(*args, **kwargs)
-        finally:
-            shutdown_cache_session()
-        return response
-    return __wrapper__
+
+def format_filters(filters):
+    """ Convert filters to a string. """
+    return "; ".join(
+        "%s: %g,%g" % (key, vmin, vmax)
+        for key, (vmin, vmax) in filters.iteritems()
+    )
+
+
+class RejectAll(Filter):
+    """ Filter rejecting all records. """
+
+    @property
+    def required_variables(self):
+        return ()
+
+    def filter(self, dataset, index=None):
+        return empty(0, dtype='int64')
+
+    def __str__(self):
+        return "RejectAll()"
