+++ vires/vires/processes/util/input_parsers.py
@@ -2,9 +2,7 @@
 #
 # Process Utilities - Input Parsers
 #
-# Project: VirES
 # Authors: Martin Paces <martin.paces@eox.at>
-#
 #-------------------------------------------------------------------------------
 # Copyright (C) 2016 EOX IT Services GmbH
 #
@@ -26,7 +24,8 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 #-------------------------------------------------------------------------------
-# pylint: disable=missing-docstring,import-error
+# pylint: disable=missing-docstring,too-many-branches,unused-argument
+
 import re
 from collections import OrderedDict
 from eoxmagmod import load_model_shc
@@ -39,14 +38,13 @@ from vires.parsers.model_list_parser import get_model_list_parser
 from vires.parsers.model_list_lexer import get_model_list_lexer
 from vires.parsers.model_expression_parser import get_model_expression_parser
 from vires.parsers.model_expression_lexer import get_model_expression_lexer
-from .time_series_product import ProductTimeSeries
-from .time_series_custom_data import CustomDatasetTimeSeries
-from .model_magmod import SourceMagneticModel, ComposedMagneticModel
+from .time_series import ProductTimeSeries, CustomDatasetTimeSeries
+from .models import SourceMagneticModel, ComposedMagneticModel
 from .filters import ScalarRangeFilter, VectorComponentRangeFilter
 
 
 RE_FILTER_NAME = re.compile(r'(^[^[]+)(?:\[([0-9])\])?$')
-RE_RESIDUAL_VARIABLE = re.compile(r'(.+)_res([ABC])([ABC])')
+RE_SUBTRACTED_VARIABLE = re.compile(r'(.+)_(?:res|diff)([ABC])([ABC])')
 
 
 def parse_style(input_id, style):
@@ -337,10 +335,10 @@ def parse_variables(input_id, variables_strings):
     ] if variables_strings else []
 
 
-def get_residual_variables(variables):
-    """ Extract residual variables from a list of all variables. """
+def get_subtracted_variables(variables):
+    """ Extract subtracted variables from a list of all variables. """
     return [
         (variable, match.groups()) for variable, match in (
-            (var, RE_RESIDUAL_VARIABLE.match(var)) for var in variables
+            (var, RE_SUBTRACTED_VARIABLE.match(var)) for var in variables
         ) if match
     ]
