+++ vires/vires/processes/util/models/magnetic_coordinates.py
@@ -2,9 +2,7 @@
 #
 # Data Source - quasi-dipole coordinates and magnetic local time
 #
-# Project: VirES
 # Authors: Martin Paces <martin.paces@eox.at>
-#
 #-------------------------------------------------------------------------------
 # Copyright (C) 2016 EOX IT Services GmbH
 #
@@ -26,7 +24,7 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 #-------------------------------------------------------------------------------
-#pylint: disable=too-many-locals
+#pylint: disable=too-many-locals,missing-docstring
 
 from logging import getLogger, LoggerAdapter
 from numpy import stack
@@ -36,7 +34,7 @@ from eoxmagmod import (
 from vires.util import include, unique
 from vires.cdf_util import cdf_rawtime_to_mjd2000, CDF_DOUBLE_TYPE
 from vires.dataset import Dataset
-from .model import Model
+from .base import Model
 
 
 class QuasiDipoleCoordinates(Model):
