+++ vires/vires/processes/util/png_output.py
@@ -2,9 +2,7 @@
 #
 # Process utilities - PNG output
 #
-# Project: VirES
 # Authors: Martin Paces <martin.paces@eox.at>
-#
 #-------------------------------------------------------------------------------
 # Copyright (C) 2016 EOX IT Services GmbH
 #
